moved flow-types to root of project
