+++ flow-types/Authority.js
