+++ flow-types/BikeRentalStation.js
