+++ flow-types/JourneyPattern.js
