+++ flow-types/Operator.js
