+++ flow-types/Situation.js
