+++ flow-types/TripPattern.js
